blog_mustafa_murat_coskun/requirements.txt
asgiref==3.2.7
Django==3.0.4
django-ckeditor==5.9.0
django-cleanup==4.0.0
django-crispy-forms==1.9.0
django-js-asset==1.2.2
pytz==2019.3
sqlparse==0.3.1
